hardware/relay_manager.py
"""
===============================================================================
  File        : relay_manager.py
  Project     : Industrial Remote Relay IoT Node
  Description : Manages the relays

  Author      : Hamas
  Created on  : 2025-07-21
  Version     : 1.0.0

  Dependencies:
    - rpi-lgpio

  Internal Modules:
    - utils.logger

  Notes:
    - Make sure pins are set up in /config/constants.py

===============================================================================
"""
# hardware/relay_manager.py

import threading
import time
from   utils.logger import getLogger

try:
    import RPi.GPIO as GPIO
    ON_RPI = True
except (ImportError, RuntimeError):
    ON_RPI = False

class RelayManager:
    POLL_INTERVAL = 0.1

    def __init__(self, config, event_queue=None):
        """
        Initializes the RelayManager.

        """

    def restartGSM(self):
        """
        Restarts the GSM module by toggling the GSM pin.

        """
        
    def _startPolling(self):
        """
        Starts a background thread to poll external relay inputs.

        """

        def poll_loop():
            """
            Internal loop for polling external relay inputs when edge detection fails.

            """

    def getRelayStatus(self):
        """
        Returns a dictionary of relay states, where the keys are the relay numbers and
        the values are the current states of the relays (True for ON, False for OFF).

        """
    
    def getExternalStatus(self):
        """
        Returns a dictionary of external relay states, where the keys are the relay numbers and
        the values are the current states of the relays (True for ON, False for OFF).

        """
    
    def _handleInputChange(self, pin):
        """
        Handles changes in external relay inputs when an edge is detected.

        """

    def setRelayState(self, relayNum, state: bool):
        """
        Sets the state of a relay.

        """

    def _handleExternalChange(self, relayNum, state):
        """
        Handles changes in external relay inputs when polled.

        """

    def close(self):
        """
        Performs cleanup operations on GPIO resources.

        """